# refactor migrations

+++ database/migrations/2024_04_06_071319_create_orders_table.php
@@ -14,11 +14,11 @@ return new class extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id')->nullable();
-            $table->float('total');
+            $table->decimal('total', 10, 2);
             $table->string('order_type');
             $table->string('status')->default("Chờ xác nhận");
             $table->timestamps();
-            
+
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
         });
     }
